+++ CHANGELOG.md
@@ -21,6 +21,14 @@ Change categories:
 
 ### Changed
 
+- base: Rename symbols:
+  - `encode_base_multipart_receive_req()` to
+    `encode_pldm_base_multipart_receive_req()`
+  - `decode_base_multipart_receive_resp()` to
+    `decode_pldm_base_multipart_receive_resp()`
+  - `struct pldm_multipart_receive_resp` to
+    `struct pldm_base_multipart_receive_resp`
+
 ### Deprecated
 
 ### Removed